requirements.txt
flask
flask-cors
python-dotenv
groq
PyPDF2
python-docx
requests
gunicorn
